Move things around to reduce symbols in main package
