+++ command/config_test.go
@@ -1,4 +1,4 @@
-package main
+package command
 
 import (
 	"encoding/json"
