+++ command/oauth2_test.go
@@ -1,4 +1,4 @@
-package main
+package command
 
 import (
 	"net/http"
