+++ command/roles.go
@@ -1,4 +1,4 @@
-package main
+package command
 
 import (
 	"github.com/spf13/cobra"
